PROXY-9: refactor project structure
