+++ src/config/environment.js
@@ -1,0 +1,9 @@
+require('dotenv').config();
+
+const { ASTEROIDS_FEED_URL, API_KEY, PORT } = process.env;
+
+module.exports = {
+    ASTEROIDS_FEED_URL,
+    API_KEY,
+    PORT,
+};
