+++ src/config/index.js
@@ -1,0 +1,7 @@
+const { ASTEROIDS_FEED_URL, API_KEY, PORT } = require('./environment');
+
+module.exports = {
+    asteroidsFeedUrl: ASTEROIDS_FEED_URL,
+    apiKey: API_KEY,
+    port: PORT,
+};
