+++ src/controllers/index.js
@@ -1,0 +1,5 @@
+const asteroidController = require('./asteroid-controller');
+
+module.exports = {
+    asteroidController,
+};
