--- src/helpers/math.js
@@ -1,5 +1,0 @@
-const getAverageValue = (array) => array.reduce((a, b) => a + b, 0) / array.length;
-
-module.exports = {
-    getAverageValue,
-};
