+++ src/routes/index.js
@@ -1,0 +1,5 @@
+const { meteorRouter } = require('./meteor-routes');
+
+module.exports = {
+    meteorRouter,
+};
