+++ src/routes/meteor-routes.js
@@ -1,0 +1,11 @@
+const express = require('express');
+const { asteroidController } = require('../controllers');
+
+const router = express.Router();
+
+// Get asteroids data
+router.get('/', asteroidController.getAsteroids);
+
+module.exports = {
+    meteorRouter: router,
+};
