--- src/routes/meteors.js
@@ -1,9 +1,0 @@
-const express = require('express');
-const { getAsteroidsData } = require('../controllers/asteroids');
-
-const router = express.Router();
-
-// Get asteroids data
-router.get('/', getAsteroidsData);
-
-module.exports = router;
