+++ src/services/index.js
@@ -1,0 +1,5 @@
+const asteroidService = require('./asteroid-service');
+
+module.exports = {
+    asteroidService,
+};
