+++ src/utils/index.js
@@ -1,0 +1,5 @@
+const { mapAsteroidsData } = require('./asteroid-mapper');
+
+module.exports = {
+    mapAsteroidsData,
+};
